+++ app/search/page.tsx
@@ -1,7 +1,7 @@
 import {H1} from "@components/elements/headers";
 import {getConfigPage} from "@lib/gql/gql-queries";
 import {StanfordBasicSiteSetting} from "@lib/gql/__generated__/drupal.d";
-import AlgoliaSearch from "./algolia-search";
+import AlgoliaSiteSearch from "@components/algolia/algolia-site-search";
 import {IndexUiState} from "instantsearch.js/es/types/ui-state";
 
 // https://nextjs.org/docs/app/api-reference/file-conventions/route-segment-config
@@ -27,7 +27,7 @@ const Page = async ({searchParams}: { searchParams?: { [_key: string]: string } 
       <H1>Search</H1>
 
       {(siteSettingsConfig?.suSiteAlgoliaId && siteSettingsConfig?.suSiteAlgoliaIndex && siteSettingsConfig?.suSiteAlgoliaSearch) &&
-        <AlgoliaSearch
+        <AlgoliaSiteSearch
           appId={siteSettingsConfig.suSiteAlgoliaId}
           searchIndex={siteSettingsConfig.suSiteAlgoliaIndex}
           searchApiKey={siteSettingsConfig.suSiteAlgoliaSearch}
