+++ package.json
@@ -16,15 +16,15 @@
     "@formkit/auto-animate": "^0.8.2",
     "@heroicons/react": "^2.1.3",
     "@js-temporal/polyfill": "^0.4.4",
-    "@mui/base": "^5.0.0-beta.43",
+    "@mui/base": "^5.0.0-beta.46",
     "@next/third-parties": "^14.2.3",
     "@tailwindcss/container-queries": "^0.1.1",
-    "@types/node": "^20.12.8",
-    "@types/react": "^18.3.1",
+    "@types/node": "^20.12.12",
+    "@types/react": "^18.3.2",
     "@types/react-dom": "^18.3.0",
     "algoliasearch": "^4.23.3",
     "autoprefixer": "^10.4.19",
-    "axios": "^1.6.8",
+    "axios": "^1.7.2",
     "clsx": "^2.1.1",
     "decanter": "^7.3.0",
     "drupal-jsonapi-params": "^2.3.1",
@@ -42,11 +42,11 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-focus-lock": "^2.12.1",
-    "react-instantsearch": "^7.7.3",
-    "react-instantsearch-nextjs": "^0.2.2",
+    "react-instantsearch": "^7.9.0",
+    "react-instantsearch-nextjs": "^0.2.5",
     "react-slick": "^0.30.2",
     "react-tiny-oembed": "^1.1.0",
-    "sharp": "^0.33.3",
+    "sharp": "^0.33.4",
     "tailwind-merge": "^2.3.0",
     "tailwindcss": "^3.4.3",
     "typescript": "^5.4.5",
@@ -57,15 +57,15 @@
     "@graphql-codegen/cli": "^5.0.2",
     "@graphql-codegen/import-types-preset": "^3.0.0",
     "@graphql-codegen/typescript-graphql-request": "^6.2.0",
-    "@graphql-codegen/typescript-operations": "^4.2.0",
+    "@graphql-codegen/typescript-operations": "^4.2.1",
     "@next/bundle-analyzer": "^14.2.3",
-    "@storybook/addon-essentials": "^8.0.9",
-    "@storybook/addon-interactions": "^8.0.9",
-    "@storybook/addon-links": "^8.0.9",
+    "@storybook/addon-essentials": "^8.1.2",
+    "@storybook/addon-interactions": "^8.1.2",
+    "@storybook/addon-links": "^8.1.2",
     "@storybook/addon-styling": "^1.3.7",
-    "@storybook/blocks": "^8.0.9",
-    "@storybook/nextjs": "^8.0.9",
-    "@storybook/react": "^8.0.9",
+    "@storybook/blocks": "^8.1.2",
+    "@storybook/nextjs": "^8.1.2",
+    "@storybook/react": "^8.1.2",
     "@storybook/testing-library": "^0.2.2",
     "@types/react-slick": "^0.23.13",
     "concurrently": "^8.2.2",
@@ -74,7 +74,7 @@
     "eslint-plugin-storybook": "^0.8.0",
     "eslint-plugin-unused-imports": "^3.2.0",
     "react-docgen": "^7.0.3",
-    "storybook": "^8.0.9",
+    "storybook": "^8.1.2",
     "tsconfig-paths-webpack-plugin": "^4.1.0"
   },
   "packageManager": "yarn@4.1.1"
