+++ src/lib/hooks/useFavorites.tsx
@@ -36,11 +36,7 @@ const useFavorites = (): {
   const toggleFav = useCallback(
     (uuid: string, title: string, path: string, units: number) => {
       const isFav = favs.some((fav) => fav.uuid === uuid);
-      if (isFav) {
-        removeFav(uuid);
-      } else {
-        addFav(uuid, title, path, units);
-      }
+      isFav ? removeFav(uuid) : addFav(uuid, title, path, units);
     },
     [favs, addFav, removeFav]
   );
